feat: Decompose FilesService into single-concern services
