+++ openspec/changes/archive/2026-07-16-split-files-service/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-07-16
